New userscript
Auto-hides clicked reports, in addition to significantly increasing color contrast between visited and unvisited to make it faster finding unchecked reports, in addition to making it possible to open a few reports of a massive batch without losing track of which are opened and which are not. 

Unfortunately, this is slightly limited as it doesn't hide already visited ones due to limitations enforced by most browsers.